tools/look_accessibility/training/requirements-training.txt
numpy>=1.24
pandas>=2.0
Pillow>=10.0
scikit-learn>=1.4
joblib>=1.3
opencv-python>=4.9
